shift task test into ui-interaction test
